Map panel / Add min and max value.
Currently colors are scaled between min and max in the
datasets. This make comparison of maps hards due to
different color scales.

Add a min and max value to force the scale to be in between
those bounds. Values outside the bounds are ignored.

By default no min/max value set.